feature: create modularity, improve documentation, prepare for testing
